[posixtime] Queue additional signal to deal with race condition
The logic of this change is:

If we get an ESRCH in timerRoutine stopping sequence, we know the
process has been reaped and the underlying POSIX timer has been
disarmed.  We queue an additional signal into the RT signal queue.  When
the demuxer see this additional signal, it can ensure that all signals
of this timer has been dealt.

Close #14.